Mark package plugins as being available in 5.5 (albeit with the preliminary API) so that package manifests don't need to set 5.6 just in order to include plugins
